Move file HubTimer.java, change HubTimer to extend SubsystemBase for internal periodic method,  and create LED code getter based on hub status
