+++ src/main/java/org/team157/robot/BuildConstants.java
@@ -5,12 +5,12 @@ public final class BuildConstants {
     public static final String MAVEN_GROUP = "";
     public static final String MAVEN_NAME = "FRC2026";
     public static final String VERSION = "unspecified";
-    public static final int GIT_REVISION = 370;
-    public static final String GIT_SHA = "ebd9d50303156e792ded3156299f467150c7e396";
-    public static final String GIT_DATE = "2026-05-19 18:30:02 EDT";
+    public static final int GIT_REVISION = 371;
+    public static final String GIT_SHA = "a58015065bab65ec2dd2b2bcfab9bd0a3a94c82a";
+    public static final String GIT_DATE = "2026-05-19 19:20:39 EDT";
     public static final String GIT_BRANCH = "feature/hub-timer";
-    public static final String BUILD_DATE = "2026-05-19 19:02:47 EDT";
-    public static final long BUILD_UNIX_TIME = 1779231767167L;
+    public static final String BUILD_DATE = "2026-05-19 20:25:24 EDT";
+    public static final long BUILD_UNIX_TIME = 1779236724388L;
     public static final int DIRTY = 1;
 
     private BuildConstants() {}
